+++ zetaclient/tss_signer.go
@@ -183,7 +183,7 @@ func (tss *TSS) Pubkey() []byte {
 // Sign signs a digest
 // digest should be Hashes of some data
 // NOTE: Specify optionalPubkey to use a different pubkey than the current pubkey set during keygen
-func (tss *TSS) Sign(digest []byte, height uint64, nonce uint64, chain *common.Chain, optionalPubKey string) ([65]byte, error) {
+func (tss *TSS) Sign(digest []byte, height uint64, _ uint64, _ *common.Chain, optionalPubKey string) ([65]byte, error) {
 	H := digest
 	log.Debug().Msgf("hash of digest is %s", H)
 
@@ -200,14 +200,16 @@ func (tss *TSS) Sign(digest []byte, height uint64, nonce uint64, chain *common.C
 	if ksRes.Status == thorcommon.Fail {
 		log.Warn().Msgf("keysign status FAIL posting blame to core, blaming node(s): %#v", ksRes.Blame.BlameNodes)
 
-		digest := hex.EncodeToString(digest)
-		index := observertypes.GetBlameIndex(chain.ChainId, nonce, digest, height)
+		//digest := hex.EncodeToString(digest)
+		//index := observertypes.GetBlameIndex(chain.ChainId, nonce, digest, height)
+		//
+		//zetaHash, err := tss.CoreBridge.PostBlameData(&ksRes.Blame, chain.ChainId, index)
+		//if err != nil {
+		//	log.Error().Err(err).Msg("error sending blame data to core")
+		//	return [65]byte{}, err
+		//}
 
-		zetaHash, err := tss.CoreBridge.PostBlameData(&ksRes.Blame, chain.ChainId, index)
-		if err != nil {
-			log.Error().Err(err).Msg("error sending blame data to core")
-			return [65]byte{}, err
-		}
+		//log.Info().Msgf("keysign posted blame data tx hash: %s", zetaHash)
 
 		// Increment Blame counter
 		for _, node := range ksRes.Blame.BlameNodes {
@@ -218,8 +220,6 @@ func (tss *TSS) Sign(digest []byte, height uint64, nonce uint64, chain *common.C
 			}
 			counter.Inc()
 		}
-
-		log.Info().Msgf("keysign posted blame data tx hash: %s", zetaHash)
 	}
 	signature := ksRes.Signatures
 
@@ -257,7 +257,7 @@ func (tss *TSS) Sign(digest []byte, height uint64, nonce uint64, chain *common.C
 
 // SignBatch is hash of some data
 // digest should be batch of hashes of some data
-func (tss *TSS) SignBatch(digests [][]byte, height uint64, nonce uint64, chain *common.Chain) ([][65]byte, error) {
+func (tss *TSS) SignBatch(digests [][]byte, height uint64, _ uint64, _ *common.Chain) ([][65]byte, error) {
 	tssPubkey := tss.CurrentPubkey
 	digestBase64 := make([]string, len(digests))
 	for i, digest := range digests {
@@ -273,14 +273,16 @@ func (tss *TSS) SignBatch(digests [][]byte, height uint64, nonce uint64, chain *
 
 	if ksRes.Status == thorcommon.Fail {
 		log.Warn().Msg("keysign status FAIL posting blame to core")
-		digest := combineDigests(digestBase64)
-		index := observertypes.GetBlameIndex(chain.ChainId, nonce, hex.EncodeToString(digest), height)
+		//digest := combineDigests(digestBase64)
+		//index := observertypes.GetBlameIndex(chain.ChainId, nonce, hex.EncodeToString(digest), height)
 
-		zetaHash, err := tss.CoreBridge.PostBlameData(&ksRes.Blame, chain.ChainId, index)
-		if err != nil {
-			log.Error().Err(err).Msg("error sending blame data to core")
-			return [][65]byte{}, err
-		}
+		//zetaHash, err := tss.CoreBridge.PostBlameData(&ksRes.Blame, chain.ChainId, index)
+		//if err != nil {
+		//	log.Error().Err(err).Msg("error sending blame data to core")
+		//	return [][65]byte{}, err
+		//}
+		//
+		//log.Info().Msgf("keysign posted blame data tx hash: %s", zetaHash)
 
 		// Increment Blame counter
 		for _, node := range ksRes.Blame.BlameNodes {
@@ -291,8 +293,6 @@ func (tss *TSS) SignBatch(digests [][]byte, height uint64, nonce uint64, chain *
 			}
 			counter.Inc()
 		}
-
-		log.Info().Msgf("keysign posted blame data tx hash: %s", zetaHash)
 	}
 
 	signatures := ksRes.Signatures
@@ -580,7 +580,7 @@ func verifySignature(tssPubkey string, signature []keysign.Signature, H []byte) 
 	return bytes.Equal(pubkey.Bytes(), compressedPubkey)
 }
 
-func combineDigests(digestList []string) []byte {
+func CombineDigests(digestList []string) []byte {
 	digestConcat := strings.Join(digestList[:], "")
 	digestBytes := chainhash.DoubleHashH([]byte(digestConcat))
 	return digestBytes.CloneBytes()